+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- improvement: remove reset_password_token after password changed in url
+
 ## v6.4.18 2025 December 8
 
 - Fix a bug: unable to run bundle install in docker image